+++ Wabbajack.Lib/Downloaders/AbstractIPS4Downloader.cs
@@ -5,6 +5,7 @@ using System.Net;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Web;
+using HtmlAgilityPack;
 using Newtonsoft.Json;
 using Wabbajack.Common;
 using Wabbajack.Lib.Validation;
@@ -15,8 +16,8 @@ namespace Wabbajack.Lib.Downloaders
     // IPS4 is the site used by LoversLab, VectorPlexus, etc. the general mechanics of each site are the 
     // same, so we can fairly easily abstract the state.
     // Pass in the state type via TState
-    public abstract class AbstractIPS4Downloader<TDownloader, TState> : AbstractNeedsLoginDownloader, IDownloader 
-        where TState : AbstractIPS4Downloader<TDownloader, TState>.State<TDownloader>, new() 
+    public abstract class AbstractIPS4Downloader<TDownloader, TState> : AbstractNeedsLoginDownloader, IDownloader
+        where TState : AbstractIPS4Downloader<TDownloader, TState>.State<TDownloader>, new()
         where TDownloader : IDownloader
     {
         public override string SiteName { get; }
@@ -61,7 +62,7 @@ namespace Wabbajack.Lib.Downloaders
         }
 
 
-        public class State<TDownloader> : AbstractDownloadState where TDownloader : IDownloader
+        public class State<TDownloader> : AbstractDownloadState, IMetaState where TDownloader : IDownloader
         {
             public string FileID { get; set; }
             public string FileName { get; set; }
@@ -69,9 +70,11 @@ namespace Wabbajack.Lib.Downloaders
             private static bool IsHTTPS => Downloader.SiteURL.AbsolutePath.StartsWith("https://");
             private static string URLPrefix => IsHTTPS ? "https://" : "http://";
 
-            private static string Site => string.IsNullOrWhiteSpace(Downloader.SiteURL.Query)
+            public static string Site => string.IsNullOrWhiteSpace(Downloader.SiteURL.Query)
                 ? $"{URLPrefix}{Downloader.SiteURL.Host}"
                 : Downloader.SiteURL.ToString();
+
+            public static AbstractNeedsLoginDownloader Downloader => (AbstractNeedsLoginDownloader)(object)DownloadDispatcher.GetInstance<TDownloader>();
 
             public override object[] PrimaryKey
             {
@@ -100,13 +103,13 @@ namespace Wabbajack.Lib.Downloaders
 
             private async Task<Stream> ResolveDownloadStream()
             {
-                var downloader = (AbstractNeedsLoginDownloader)(object)DownloadDispatcher.GetInstance<TDownloader>();
+                //var downloader = (AbstractNeedsLoginDownloader)(object)DownloadDispatcher.GetInstance<TDownloader>();
 
                 TOP:
                 var csrfurl = FileID == null
                     ? $"{Site}/files/file/{FileName}/?do=download"
                     : $"{Site}/files/file/{FileName}/?do=download&r={FileID}";
-                var html = await downloader.AuthedClient.GetStringAsync(csrfurl);
+                var html = await Downloader.AuthedClient.GetStringAsync(csrfurl);
 
                 var pattern = new Regex("(?<=csrfKey=).*(?=[&\"\'])|(?<=csrfKey: \").*(?=[&\"\'])");
                 var matches = pattern.Matches(html).Cast<Match>();
@@ -122,10 +125,10 @@ namespace Wabbajack.Lib.Downloaders
                     : $"{Site}/files/file/{FileName}/{sep}do=download&r={FileID}&confirm=1&t=1&csrfKey={csrfKey}";
                     
 
-                var streamResult = await downloader.AuthedClient.GetAsync(url);
+                var streamResult = await Downloader.AuthedClient.GetAsync(url);
                 if (streamResult.StatusCode != HttpStatusCode.OK)
                 {
-                    Utils.ErrorThrow(new InvalidOperationException(), $"{downloader.SiteName} servers reported an error for file: {FileID}");
+                    Utils.ErrorThrow(new InvalidOperationException(), $"{Downloader.SiteName} servers reported an error for file: {FileID}");
                 }
 
                 var contentType = streamResult.Content.Headers.ContentType;
@@ -138,7 +141,7 @@ namespace Wabbajack.Lib.Downloaders
                 var secs = times.Download - times.CurrentTime;
                 for (int x = 0; x < secs; x++)
                 {
-                    Utils.Status($"Waiting for {secs} at the request of {downloader.SiteName}", Percent.FactoryPutInRange(x, secs));
+                    Utils.Status($"Waiting for {secs} at the request of {Downloader.SiteName}", Percent.FactoryPutInRange(x, secs));
                     await Task.Delay(1000);
                 }
                 streamResult.Dispose();
@@ -200,7 +203,19 @@ namespace Wabbajack.Lib.Downloaders
                 };
             }
 
-            private static AbstractNeedsLoginDownloader Downloader => (AbstractNeedsLoginDownloader)(object)DownloadDispatcher.GetInstance<TDownloader>();
+            // from IMetaState
+            public string URL => $"{Site}/files/file/{FileName}";
+            public string Name { get; set; }
+            public string Author { get; set; }
+            public string Version { get; set; }
+            public string ImageURL { get; set; }
+            public virtual bool IsNSFW { get; set; }
+            public string Description { get; set; }
+
+            public virtual async Task<bool> LoadMetaData()
+            {
+                return false;
+            }
         }
 
         protected AbstractIPS4Downloader(Uri loginUri, string encryptedKeyName, string cookieDomain) :